only delete eventteams if event is over

+++ controllers/cron_controller.py
@@ -75,11 +75,13 @@ class EventTeamUpdate(webapp.RequestHandler):
                 team_ids.add(team)
 
         # Delete EventTeams for teams who did not participate in the event
+        # Only runs if event is over
         et_keys_to_delete = set()
-        for team_id in existing_team_ids.difference(team_ids):
-            et_key_name = "{}_{}".format(event.key_name, team_id)
-            et_keys_to_delete.add(ndb.Key(EventTeam, et_key_name))
-        ndb.delete_multi(et_keys_to_delete)
+        if event.end_date < datetime.datetime.now():
+            for team_id in existing_team_ids.difference(team_ids):
+                et_key_name = "{}_{}".format(event.key_name, team_id)
+                et_keys_to_delete.add(ndb.Key(EventTeam, et_key_name))
+            ndb.delete_multi(et_keys_to_delete)
 
         # Create or update EventTeams
         teams = TeamManipulator.createOrUpdate([Team(
